Parameterized x-end.scad vertical distances and leadscrew
The distance between the leadscrew and the vertical smooth
rod is now configurable between 17mm (original Prusa i3 MK2)
and up to about 30mm for whoever might need it. This means,
that the x-end now also support the 22mm spacing of the Anet A8.

In addition, I added an option to remove the mount for the
leadscrew. If disabled, only a body for the smooth rod bearing
will be created, the mount for the TR nut will not be generated.

This is useful when using an anti-z-wobble coupling.