x-end-idler now uses parameterized x-end
The x-end-idler uses the new parameterized version of the
x-end common file. The idler is now paramterized as well,
even tough selective infill was a bit of a PITA to adapt.

You can now create and x-end-idler with and without a lead
screw and a distance between smooth rod and leadscrew from
17mm (default i3 MK2) and about 30mm.

Also updated the STL for the 17mm version to match this SCAD
file. If I didn't mess up while editing, the STL should pro-
duce an identical model as the previous version (although
it does not produce an identical STL, which wasn't to be
expected as even the original unmodified SCAD file does not
render to the same STL).