x-end-motor now uses parameterized x-end
The x-end-motor uses the new parameterized version of the
x-end common file. The motor side is now paramterized as well,
and uses the same selective infill as the idler side.

You can now create and x-end-motor with and without a
lead screw and a distance between smooth rod and leadscrew
17mm (default i3 MK2) and about 30mm. When no leadscrew
is used, the endstop mount switches to the top of the motor end,
so the bottom stays flat and can be used with an anti-z-wobble
coupling.

Also updated the STL for the 17mm version to match this SCAD
file. If I didn't mess up while editing, the STL should pro-
duce an identical model as the previous version (although
it does not produce an identical STL, which wasn't to
be expected as even the original unmodified SCAD file does not
render to the same STL).